Prefer item's 'content', if present and allowing HTML
Those stripping HTML (via the '-s' option) should be unaffected.

For those *not* stripping HTML, we now check whether posts contain 'content' of
type 'text/html': if so, we produce a multipart HTML email, containing that
'content' as the HTML part; with the 'description' as the plaintext part.
(Previously we were only including the 'description', and writing it straight
into a plaintext email, rather than being multipart).

I've found this useful in conjunction with "morss", which attaches HTML
'content' to posts by looking for the "full pages" that posts are summaries of.